Standardized Momentum Loss
Stands that can not move in stopped time no longer lose momentum in it.
Momentum is now a double and not an integer

+++ src/main/java/io/github/halffocused/diamond_is_uncraftable/capability/Stand.java
@@ -68,7 +68,7 @@ public class Stand implements ICapabilitySerializable<INBT> {
     private boolean preventUnsummon;
     private boolean preventUnsummon2;
     private boolean preventUnsummon3;
-    private int momentum = 0;
+    private double momentum = 0;
     private int restraint = 0;
     private int epitaphEffectTicker = 0;
     private double maxStandEnergy = 100;
@@ -139,7 +139,7 @@ public class Stand implements ICapabilitySerializable<INBT> {
                 nbt.putBoolean("preventUnsummon", instance.getPreventUnsummon());
                 nbt.putBoolean("preventUnsummon2", instance.getPreventUnsummon2());
                 nbt.putBoolean("preventUnsummon3", instance.getPreventUnsummon3());
-                nbt.putInt("momentum", instance.getMomentum());
+                nbt.putDouble("momentum", instance.getMomentum());
                 nbt.putInt("restraint", instance.getRestraint());
                 nbt.putInt("epitaphTicker", instance.getTimeSkipEffectTicker());
                 nbt.putDouble("currentEnergy", instance.getCurrentStandEnergy());
@@ -211,7 +211,7 @@ public class Stand implements ICapabilitySerializable<INBT> {
                 instance.preventUnsummon = compoundNBT.getBoolean("preventUnsummon");
                 instance.preventUnsummon2 = compoundNBT.getBoolean("preventUnsummon2");
                 instance.preventUnsummon3 = compoundNBT.getBoolean("preventUnsummon3");
-                instance.momentum = compoundNBT.getInt("momentum");
+                instance.momentum = compoundNBT.getDouble("momentum");
                 instance.restraint = compoundNBT.getInt("restraint");
                 instance.epitaphEffectTicker = compoundNBT.getInt("epitaphTicker");
                 instance.currentStandEnergy = compoundNBT.getDouble("currentEnergy");
@@ -359,11 +359,11 @@ public class Stand implements ICapabilitySerializable<INBT> {
         onDataUpdated();
     }
 
-    public int getMomentum() {
+    public double getMomentum() {
         return momentum;
     }
 
-    public void setMomentum(int momentumIn) {
+    public void setMomentum(double momentumIn) {
         this.momentum = momentumIn;
         onDataUpdated();
     }
